Add capsule booking costs. Add customer preferred capsule to UI

+++ Features/CustomerV2/CustomerData.cs
@@ -2,7 +2,7 @@
 {
     public string Name;
 
-    public float StayDuration;
+    public int StayDuration;
     
     public CapsuleConfiguration PreferredCapsule;
 }
